requirements.txt
flask
tensorflow
numpy
pillow
gdown
gunicorn
